git-svn-id: svn://192.168.1.10/public/CoffeeMud@2975 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/StdInnKey.java
@@ -37,7 +37,10 @@ public class StdInnKey extends StdKey implements InnKey
 			if(owner()!=null)
 				removeFromOwnerContainer();
 			if(myShopkeeper!=null)
-				myShopkeeper.addStoreInventory(this);
+			{
+				myShopkeeper.addStoreInventory(this); // makes a copy
+				destroy();
+			}
 			return false;
 		}
 		return true;
@@ -60,9 +63,14 @@ public class StdInnKey extends StdKey implements InnKey
 		super.affect(myHost,affect);
 		if(((affect.targetMinor()==Affect.TYP_GIVE)
 		||(affect.targetMinor()==Affect.TYP_SELL))
+		&&(myShopkeeper!=null)
 		&&(affect.target()==myShopkeeper)
 		&&(affect.tool()==this))
+		{
 			ExternalPlay.deleteTick(this,Host.ITEM_BOUNCEBACK);
+			myShopkeeper.addStoreInventory(this); //makes a copy
+			destroy();
+		}
 	}
 	public boolean okAffect(Environmental myHost, Affect affect)
 	{
@@ -80,11 +88,10 @@ public class StdInnKey extends StdKey implements InnKey
 			return false;
 		}
 		else
-		if((affect.sourceMinor()==Affect.TYP_BUY)
-		&&(affect.target()==this)
-		&&(affect.tool()!=null)
+		if((affect.sourceMinor()==Affect.TYP_GET)
 		&&(myShopkeeper!=null)
-		&&(affect.tool()==myShopkeeper))
+		&&(affect.tool()==myShopkeeper)
+		&&(affect.target()==this))
 			ExternalPlay.startTickDown(this,Host.ITEM_BOUNCEBACK,(int)Host.TICKS_PER_DAY);
 		return true;
 	}
